Delete PVC when node is tainted as out-of-service
This change adds logic to scan Galera pods and identify if they are running on nodes
tainted with `node.kubernetes.io/out-of-service`. If such taints are found, the
corresponding pod's PVC is deleted to allow the StatefulSet controller to re-create
the pod on a healthy node. This is particularly useful when using self-node-remediation
strategies that remove pods without handling PVC cleanup.